Open service: extract query-runtime and unify query get builders
Move query reads, loads, the drain loop, and subscriptions into query-runtime.ts
so service-runtime.ts stays focused on runtime assembly. Collapse the three
duplicate get() implementations into createQueryGet with per-surface load policies.
